+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Next
 
+- Simplified API: everything other than the initial URL or HTML is provided as options
 - Adds support for matching generating functions in Twig and as actions, updates README
 - Adds support for using an environment variable for settings API key
 - Pins api2pdf dependency to specific version (v1.1.1)